rewrite: scraper
I rewrote the scraper to scrape all information needed to create links
to jobs on the CircleCI dashboard.

I didn't find a way to link to specific tests, only to jobs those tests
belong to. I couldn't even find a separate test page in the UI, so I think
this is the best we could do.

Also, now it writes output to stdout, but if you need it in a file, you
can easily pipe the output to a file, I think this is a better approach.